(PC-12439)[API] refactor: move ubble endpoint to subscription module
