+++ api/src/pcapi/routes/native/v1/account.py
@@ -8,13 +8,10 @@ import pydantic
 
 from pcapi.connectors import api_recaptcha
 from pcapi.connectors import user_profiling
-from pcapi.connectors.beneficiaries import exceptions as beneficiaries_exceptions
 from pcapi.core.fraud import api as fraud_api
-from pcapi.core.fraud.ubble import api as ubble_fraud_api
 from pcapi.core.logging import get_or_set_correlation_id
 from pcapi.core.mails.transactional.users.delete_account import send_user_request_to_delete_account_reception_email
 from pcapi.core.subscription import api as subscription_api
-from pcapi.core.subscription.ubble import api as ubble_subscription_api
 from pcapi.core.users import api
 from pcapi.core.users import constants
 from pcapi.core.users import email as email_api
@@ -320,46 +317,3 @@ def profiling_session_id(user: User) -> serializers.UserProfilingSessionIdRespon
     """
     session_id = str(uuid.uuid4())
     return serializers.UserProfilingSessionIdResponse(sessionId=session_id)
-
-
-@blueprint.native_v1.route("/ubble_identification", methods=["POST"])
-@spectree_serialize(api=blueprint.api, response_model=serializers.IdentificationSessionResponse)
-@authenticated_user_required
-def start_identification_session(
-    user: User, body: serializers.IdentificationSessionRequest
-) -> serializers.IdentificationSessionResponse:
-
-    if user.eligibility is None:
-        raise ApiErrors(
-            {"code": "IDCHECK_NOT_ELIGIBLE", "message": "Non éligible à un crédit"},
-            status_code=400,
-        )
-
-    if not ubble_fraud_api.is_user_allowed_to_perform_ubble_check(user, user.eligibility):
-        raise ApiErrors(
-            {"code": "IDCHECK_ALREADY_PROCESSED", "message": "Une identification a déjà été traitée"},
-            status_code=400,
-        )
-
-    fraud_check = ubble_fraud_api.get_restartable_identity_checks(user)
-    if fraud_check:
-        return serializers.IdentificationSessionResponse(identificationUrl=fraud_check.source_data().identification_url)
-
-    try:
-        identification_url = ubble_subscription_api.start_ubble_workflow(user, body.redirectUrl)
-        return serializers.IdentificationSessionResponse(identificationUrl=identification_url)
-
-    except beneficiaries_exceptions.IdentificationServiceUnavailable:
-        raise ApiErrors(
-            {"code": "IDCHECK_SERVICE_UNAVAILABLE", "message": "Le service d'identification n'est pas joignable"},
-            status_code=503,
-        )
-
-    except beneficiaries_exceptions.IdentificationServiceError:
-        raise ApiErrors(
-            {
-                "code": "IDCHECK_SERVICE_ERROR",
-                "message": "Une erreur s'est produite à l'appel du service d'identification",
-            },
-            status_code=500,
-        )
